feat: note module fully typescripted
