+++ packages/api/package.json
@@ -53,6 +53,7 @@
     "@types/jest": "^26.0.0",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/mongoose": "^5.10.3",
+    "@types/mongoose-paginate": "^5.0.8",
     "@types/node": "^14.14.14",
     "@types/nodemailer": "^6.4.0",
     "@types/passport": "^1.0.4",
